Collapse experimental/utils/lora.py into pipeline_wan_lora.py
The LoRA fusion utilities were split into a separate utils module
during an earlier attempt to promote LoRA to first-class. That
promotion was reverted (LoRA stays experimental); the split survived
even though only LoRASpec was used across files. Putting everything
back in one module matches the pre-existing single-file layout and
removes a redundant import dependency.

While merging:
- Drop the unused fuse_lora_stack helper (never called in-tree).
- Make _lora_stack_cache_namespace private (the hardcoded
  'Wan2.2-I2V-LoRA-' prefix ties it to this pipeline; not a general
  utility).
- Pull the LOW_RANK_RE regex + SLOT_MAP out to module scope as
  _LOW_RANK_RE / _SLOT_MAP so they're compiled once instead of once
  per fuse_lora_state_dict call.
- Reorganize file layout: public types -> key-normalization helpers
  -> fusion API -> cache key -> pipeline class.

Public surface from pipeline_wan_lora is now:
  LoRASpec, LoRAArg, normalize_lora_arg, fuse_lora_state_dict,
  verify_fusion_changed_weights, WanPipelineI2VLora.
