+++ models/tt_dit/experimental/pipelines/pipeline_wan_svi.py
@@ -51,8 +51,7 @@ import PIL
 import torch
 from loguru import logger
 
-from models.tt_dit.experimental.pipelines.pipeline_wan_lora import WanPipelineI2VLora
-from models.tt_dit.experimental.utils.lora import LoRASpec
+from models.tt_dit.experimental.pipelines.pipeline_wan_lora import LoRASpec, WanPipelineI2VLora
 from models.tt_dit.pipelines.wan.pipeline_wan import WanPipeline
 from models.tt_dit.pipelines.wan.pipeline_wan_i2v import ImagePrompt
