+++ models/tt_dit/experimental/tests/test_pipeline_lora.py
@@ -22,8 +22,7 @@ import torch
 from loguru import logger
 
 import ttnn
-from models.tt_dit.experimental.pipelines.pipeline_wan_lora import WanPipelineI2VLora
-from models.tt_dit.experimental.utils.lora import LoRASpec
+from models.tt_dit.experimental.pipelines.pipeline_wan_lora import LoRASpec, WanPipelineI2VLora
 from models.tt_dit.pipelines.wan.pipeline_wan_i2v import ImagePrompt
 from models.tt_dit.utils.test import line_params, ring_params
